Moves Notification-related classes to a `.notification` package
