+++ library/src/main/java/com/novoda/downloadmanager/lib/NotificationCustomiser.java
@@ -2,12 +2,12 @@ package com.novoda.downloadmanager.lib;
 
 import android.support.v4.app.NotificationCompat;
 
-import com.novoda.downloadmanager.CancelledNotificationCustomiser;
-import com.novoda.downloadmanager.CompleteNotificationCustomiser;
+import com.novoda.downloadmanager.notifications.CancelledNotificationCustomiser;
+import com.novoda.downloadmanager.notifications.CompleteNotificationCustomiser;
 import com.novoda.downloadmanager.Download;
-import com.novoda.downloadmanager.DownloadingNotificationCustomiser;
-import com.novoda.downloadmanager.FailedNotificationCustomiser;
-import com.novoda.downloadmanager.QueuedNotificationCustomiser;
+import com.novoda.downloadmanager.notifications.DownloadingNotificationCustomiser;
+import com.novoda.downloadmanager.notifications.FailedNotificationCustomiser;
+import com.novoda.downloadmanager.notifications.QueuedNotificationCustomiser;
 
 class NotificationCustomiser implements QueuedNotificationCustomiser, DownloadingNotificationCustomiser,
         CompleteNotificationCustomiser, CancelledNotificationCustomiser, FailedNotificationCustomiser {
