+++ library/src/main/java/com/novoda/downloadmanager/notifications/DownloadingNotificationCustomiser.java
@@ -1,4 +1,4 @@
-package com.novoda.downloadmanager;
+package com.novoda.downloadmanager.notifications;
 
 import android.support.v4.app.NotificationCompat;
