+++ library/src/main/java/com/novoda/downloadmanager/notifications/EmptyCompleteNotificationCustomiser.java
@@ -1,6 +1,8 @@
-package com.novoda.downloadmanager;
+package com.novoda.downloadmanager.notifications;
 
 import android.support.v4.app.NotificationCompat;
+
+import com.novoda.downloadmanager.Download;
 
 public class EmptyCompleteNotificationCustomiser implements CompleteNotificationCustomiser {
     @Override
