+++ library/src/main/java/com/novoda/downloadmanager/notifications/NotificationImageRetriever.java
@@ -1,4 +1,4 @@
-package com.novoda.downloadmanager.lib;
+package com.novoda.downloadmanager.notifications;
 
 import android.graphics.Bitmap;
